+++ lib/pages/map/widgets/map_overlays/no_event_overlay.dart
@@ -34,7 +34,6 @@ class NoEventOverlayWidget extends StatelessWidget {
                             CupertinoColors.transparent, context),
                     child: Builder(builder: (context) {
                       return EventDataOverview(
-                        nextEvent: event,
                         showMap: false,
                       );
                     }),